project/api/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile project/pyproject.toml --extra api -o api/requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via project (project/pyproject.toml)
filelock==3.18.0
    # via torch
fsspec==2025.7.0
    # via torch
gunicorn==23.0.0
    # via project (project/pyproject.toml)
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
jinja2==3.1.6
    # via torch
lightning-utilities==0.14.3
    # via torchmetrics
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   pandas
    #   torchmetrics
    #   torchvision
packaging==25.0
    # via
    #   gunicorn
    #   lightning-utilities
    #   torchmetrics
pandas==2.3.1
    # via project (project/pyproject.toml)
pillow==11.3.0
    # via
    #   project (project/pyproject.toml)
    #   torchvision
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via project (project/pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
setuptools==80.9.0
    # via
    #   lightning-utilities
    #   torch
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
sympy==1.14.0
    # via torch
torch==2.7.1
    # via
    #   project (project/pyproject.toml)
    #   torchmetrics
    #   torchvision
torchmetrics==1.8.0
    # via project (project/pyproject.toml)
torchvision==0.22.1
    # via project (project/pyproject.toml)
typing-extensions==4.14.1
    # via
    #   fastapi
    #   lightning-utilities
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uvicorn==0.35.0
    # via project (project/pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
